+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/executable/impl/ops/BaseForExecOpIndexNestedLoopsJoinWithRequestOps.java
@@ -31,18 +31,8 @@ import se.liu.ida.hefquin.federation.FederationMember;
  */
 public abstract class BaseForExecOpIndexNestedLoopsJoinWithRequestOps<QueryType extends Query,
                                                                       MemberType extends FederationMember>
-              extends UnaryExecutableOpBaseWithBatching
+              extends BaseForUnaryExecOpWithCollectedInput
 {
-	// Since this algorithm processes the input solution mappings
-	// in parallel, we should use an input block size with which
-	// we can leverage this parallelism. However, I am not sure
-	// yet what a good value is; it probably depends on various
-	// factors, including the load on the federation member and
-	// the degree of parallelism in the FederationAccessManager.
-	// Notice that this number is essentially the number of
-	// requests issued in parallel (to the same endpoint!).
-	public final static int DEFAULT_BATCH_SIZE = 10;
-
 	protected final QueryType query;
 	protected final MemberType fm;
 	protected final boolean useOuterJoinSemantics;
@@ -55,10 +45,10 @@ public abstract class BaseForExecOpIndexNestedLoopsJoinWithRequestOps<QueryType 
 	protected BaseForExecOpIndexNestedLoopsJoinWithRequestOps( final QueryType query,
 	                                                           final MemberType fm,
 	                                                           final boolean useOuterJoinSemantics,
-	                                                           final int batchSize,
+	                                                           final int minimumInputBlockSize,
 	                                                           final boolean collectExceptions,
 	                                                           final QueryPlanningInfo qpInfo ) {
-		super(batchSize, collectExceptions, qpInfo);
+		super(minimumInputBlockSize, collectExceptions, qpInfo);
 
 		assert query != null;
 		assert fm != null;
@@ -68,16 +58,8 @@ public abstract class BaseForExecOpIndexNestedLoopsJoinWithRequestOps<QueryType 
 		this.useOuterJoinSemantics = useOuterJoinSemantics;
 	}
 
-	protected BaseForExecOpIndexNestedLoopsJoinWithRequestOps( final QueryType query,
-	                                                           final MemberType fm,
-	                                                           final boolean useOuterJoinSemantics,
-	                                                           final boolean collectExceptions,
-	                                                           final QueryPlanningInfo qpInfo ) {
-		this(query, fm, useOuterJoinSemantics, DEFAULT_BATCH_SIZE, collectExceptions, qpInfo);
-	}
-
 	@Override
-	protected void _processBatch( final List<SolutionMapping> input,
+	protected void _processCollectedInput( final List<SolutionMapping> input,
 	                              final IntermediateResultElementSink sink,
 	                              final ExecutionContext execCxt )
 			throws ExecOpExecutionException
@@ -196,7 +178,7 @@ public abstract class BaseForExecOpIndexNestedLoopsJoinWithRequestOps<QueryType 
 			throws ExecOpExecutionException
 	{
 		if ( input != null && ! input.isEmpty() ) {
-			_processBatch(input, sink, execCxt);
+			_processCollectedInput(input, sink, execCxt);
 		}
 	}