+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/executable/impl/ops/ExecOpLookupJoinViaWrapperWithParamVars.java
@@ -6,6 +6,8 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ExecutionException;
 import java.util.function.Consumer;
 
@@ -33,7 +35,7 @@ import se.liu.ida.hefquin.federation.members.WrappedRESTEndpoint;
 import se.liu.ida.hefquin.federation.members.WrappedRESTEndpoint.DataConversionException;
 
 public class ExecOpLookupJoinViaWrapperWithParamVars
-       extends UnaryExecutableOpBaseWithBatching
+       extends BaseForUnaryExecOpWithCollectedInput
 {
 	// Since this algorithm processes the input solution mappings
 	// in parallel, we should use an input block size with which
@@ -43,13 +45,13 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 	// the degree of parallelism in the FederationAccessManager.
 	// Notice that this number is essentially the number of
 	// requests issued in parallel (to the same endpoint!).
-	public final static int DEFAULT_BATCH_SIZE = 5;
+	public final static int DEFAULT_INPUT_BLOCK_SIZE = 5;
 
 	protected final SPARQLGraphPattern pattern;
 	protected final Map<String,Var> paramVars;
 	protected final WrappedRESTEndpoint fm;
 
-	protected final Map<Map<String,Node>, List<SolutionMapping>> cache = new HashMap<>();
+	protected final ConcurrentMap<Map<String,Node>, List<SolutionMapping>> cache = new ConcurrentHashMap<>();
 
 	// statistics
 	private long numberOfRequestsIssued = 0L;
@@ -65,17 +67,17 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 	                                                final boolean collectExceptions,
 	                                                final QueryPlanningInfo qpInfo ) {
 		this( pattern, paramVars, fm,
-		      DEFAULT_BATCH_SIZE,
+		      DEFAULT_INPUT_BLOCK_SIZE,
 		      collectExceptions, qpInfo );
 	}
 
 	public ExecOpLookupJoinViaWrapperWithParamVars( final SPARQLGraphPattern pattern,
 	                                                final Map<String,Var> paramVars,
 	                                                final WrappedRESTEndpoint fm,
-	                                                final int batchSize,
+	                                                final int minimumInputBlockSize,
 	                                                final boolean collectExceptions,
 	                                                final QueryPlanningInfo qpInfo ) {
-		super(batchSize, collectExceptions, qpInfo);
+		super(minimumInputBlockSize, collectExceptions, qpInfo);
 
 		assert pattern   != null;
 		assert paramVars != null;
@@ -90,7 +92,7 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 	}
 
 	@Override
-	protected void _processBatch( final List<SolutionMapping> input,
+	protected void _processCollectedInput( final List<SolutionMapping> input,
 	                              final IntermediateResultElementSink sink,
 	                              final ExecutionContext execCxt )
 			throws ExecOpExecutionException
@@ -151,7 +153,8 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 			numberOfRequestsIssued++;
 
 			// attach the processing of the response obtained for the request
-			final MyResponseProcessor respProc = new MyResponseProcessor( entry.getValue(), sink, this );
+			final MyResponseProcessor respProc = new MyResponseProcessor(
+					entry.getKey(),entry.getValue(), sink, this );
 			futures[i++] = f.thenAccept(respProc);
 		}
 
@@ -174,7 +177,7 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 			throws ExecOpExecutionException
 	{
 		if ( input != null && ! input.isEmpty() ) {
-			_processBatch(input, sink, execCxt);
+			_processCollectedInput(input, sink, execCxt);
 		}
 	}
 
@@ -239,13 +242,16 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 
 	protected class MyResponseProcessor implements Consumer<StringResponse>
 	{
+		protected final Map<String,Node> paramValues;
 		protected final List<SolutionMapping> solmaps;
 		protected final IntermediateResultElementSink sink;
 		protected final ExecutableOperator op;
 
-		public MyResponseProcessor( final List<SolutionMapping> solmaps,
+		public MyResponseProcessor( final Map<String,Node> paramValues,
+		                            final List<SolutionMapping> solmaps,
 		                            final IntermediateResultElementSink sink,
 		                            final ExecutableOperator op ) {
+			this.paramValues = paramValues;
 			this.solmaps = solmaps;
 			this.sink = sink;
 			this.op = op;
@@ -265,9 +271,9 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 				return;
 			}
 
-			final Iterable<SolutionMapping> resutingSolMaps;
+			final List<SolutionMapping> resultingSolMaps;
 			try {
-				resutingSolMaps = fm.evaluatePatternOverRDFView(pattern, respData);
+				resultingSolMaps = fm.evaluatePatternOverRDFView(pattern, respData);
 			}
 			catch ( final DataConversionException e ) {
 				numberOfDataConversionExceptions++;
@@ -276,8 +282,12 @@ public class ExecOpLookupJoinViaWrapperWithParamVars
 				return;
 			}
 
+			join(resultingSolMaps, solmaps, sink);
 
-			join(resutingSolMaps, solmaps, sink);
+			final List<SolutionMapping> x = cache.putIfAbsent( paramValues,
+			                                                   resultingSolMaps );
+			if ( x != null )
+				throw new IllegalStateException("Overwriting in the cache.");
 
 			final long time2 = System.currentTimeMillis();